+++ twitch-spy-frontend/src/backend/socket.ts
@@ -1,0 +1,1 @@
+export {};
